requirements.txt
# requirements.txt
schedule==1.2.1
selenium==4.18.1
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
tqdm==4.66.2
python-dotenv==1.0.1 # For loading .env file